refactor: Change reviews type from double to int in ProductModel and ProductInfoSection

+++ lib/presentation/screens/product/widgets/product_info_section.dart
@@ -6,7 +6,7 @@ class ProductInfoSection extends StatelessWidget {
   final double price;
   final String description;
   final double rating;
-  final double reviews;
+  final int reviews;
 
   const ProductInfoSection({
     super.key,
